requirements.txt
# VaultZero v2.0 Requirements
# Core dependencies from v1.0 + new LangGraph additions

# Streamlit for UI
streamlit==1.31.0

# LangGraph & LangChain (NEW for v2.0)
langgraph==0.2.53
langchain==0.3.17
langchain-core==0.3.32
langchain-anthropic==0.3.0

# Anthropic Claude API
anthropic==0.42.0

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.2.0

# Vector DB and embeddings (for RAG if needed)
chromadb==0.4.22

# HTTP requests
requests==2.31.0
aiohttp==3.9.3

# Data handling
pydantic==2.5.3
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.1
pathlib==1.0.1

# Cloud deployment
gunicorn==21.2.0

# Logging and monitoring
loguru==0.7.2
